Send the correct Content-Length for "foo".
We used to read "C: foo\n" and HTTP::Request would parse that, find that
it didn't match the declared "Content-Length: 3", and complain about it.
Now we remove the extra LF before feeding it to parse, and all is well.

But what if there is really an LF at the end of the request body? We try
to handle that too (with an ugly extra blank line for chomp to consume),
but I'm not certain it does the right thing. More testing needed.