Set the Host: to our own address on proxied requests.
We act as a transparent *server-side* proxy, i.e., the client talks to
our address directly, but doesn't know that we're proxying requests to
a backend server. The client will thus send us requests with relative
URIs and our own address in Host. We ensure that the Host is set that
way because the origin server uses it to generate redirects and links
in JavaScript. At the same time, we must continue to modify the URI to
refer to the origin server, as before.

These requirements seem likely to vary widely, so some configurability
will be needed eventually.